refactor: layout 패키징 분리
