+++ newsstand/src/layouts/header/Header.js
